Code cleanup and initial restructuring
