+++ documentation/requirements.txt
@@ -11,3 +11,5 @@ Werkzeug
 requests
 sphinx_rtd_theme
 lark-parser
+ws4py
+pyaudio
