--- stream/sender.py
@@ -1,10 +1,0 @@
-import sys
-
-sys.path.append("../")
-import multiprocessing
-
-import Queue
-
-
-q = Queue.Queue
-q.put(item="TESTING")
